Re-organize code to separate responsibilities
